feat(beer): Add Caloric Information to Beer
Add the caloric information about a beer.

I also went ahead and commented out the weird routing stuff that makes
it so that I can't access it from localhost. At some point, I'm going to
dive in and understand what's going on there so that I can actually
publish to Heroku.