fix(routing): Add Better Routes
Remove some of the fanciness caused by the `api` subdomain that wouldn't
allow me to test on either local or Heroku. Now the api is accessed at a
sensible `/api/v1`. Also changed the regular constraint to include any
subdomain that _isn't_ matched by the other constraints. That allows the
application to work on Heroku.