test(calories): Add Calories To All Beer Tests
I do NOT want to have to do this. I want calories to work just like ABV
does. If it's included, dandy. If not, that's fine too. Forcing every
single beer to have caloric information just isn't going to work.
However, for now, I've added calories to all of the creations in the
tests so that they pass.

I've got calories set up just like ABV. I'm not sure what the problem
is.